FIX async initialization of singletons
After adding some basic jest tests a few issues popped up:

1. Using React Hooks in tests doesn't work because hooks cannot be called
outside of components. This means our use*.ts functions can't be used to pass
around singleton instances outside of components.

2. Manually exposing instances in the singleton.ts module does not allow a
different datasource configuration in the tests than we use in the main
application.

3. The data source object was not being initialized correctly because it needs
to be initialized asynchronously. Previously we had been trying to initialize
the data source within the DbService constructor; however, constructors cannot
be reliably used to start asynchronous operations.

This commit fixes these problems by changing the singleton.ts module to export a
single async function that resolves to an object containing all our singletons
instances. This function accepts a data source as a parameter so that a test
configuration can be defined and is memoized to ensure that we only get one
instance of the result per datasource.